requirements.txt
PySide6-Essentials==6.8.1
shiboken6==6.8.1
requests==2.32.3
python-dotenv==1.0.1
matplotlib>=3.8.0
networkx>=3.2
scipy>=1.11.0
numpy>=1.24.0
